requirements.txt
python-dotenv
schedule
Flask
jsonify